refactor: extract Provider_Role from Provider_Group
- Create src/Provider_Role.php with all role-mapping DB/UI/form logic
  (CommonDBRelation, getTypeName, tab display, executeFormAction,
   cleanDBonPurge, getMappingsForProvider, getRoleMappingsForClaims,
   getConfiguredGlpiGroups, removeDynamicGroupsForRole, ensureGroupMapping,
   getDynamicGroupsTable)
- Refactor src/Provider_Group.php to a pure utility class with only
  group-sync logic (syncGroups, extractUserGroupsFromResource,
  normalizeGroupClaimValue); all Provider_Role calls updated
- Update front/provider_group.form.php to use Provider_Role
- Update src/Provider.php: addStandardTab, getTable, removeDynamicGroupsForRole
  now reference Provider_Role; syncGroups still calls Provider_Group
- Update setup.php: register Provider_Role instead of Provider_Group
- Update hook.php: use Provider_Role for table name lookups

Agent-Logs-Url: https://github.com/eduardomozart/glpi-singlesignon/sessions/a9e435da-e121-4b3a-9e6d-99adc4669660
